Adds a json endpoint for listing tags. Shared base controller for API controllers.
