+++ app/controllers/api/stories_controller.rb
@@ -1,8 +1,5 @@
 module Api
-  class StoriesController < ActionController::Base
-    before_action :find_user_from_rss_token
-    before_action :require_logged_in_user
-
+  class StoriesController < BaseController
     def create
       @story = Story.new(story_params)
       @story.user_id = @user.id
@@ -13,22 +10,9 @@ module Api
           return
         end
       end
-      render status: 422
+      render json: {}, status: 422
     end
 
-    def find_user_from_rss_token
-      if !@user && request.headers["Authorization"].to_s.present?
-        @user = User.where(:rss_token => request.headers["Authorization"].to_s).first
-      end
-    end
-
-    def require_logged_in_user
-      if @user
-        true
-      else
-        render json: {}, status: 401
-      end
-    end
 
     def story_params
       params.require(:story).permit(
